update for saving PlayingInfo

+++ app/src/main/java/com/github/lilinsong3/m3player/data/model/PlayingInfoModel.kt
@@ -1,7 +1,7 @@
 package com.github.lilinsong3.m3player.data.model
 
 data class PlayingInfoModel(
-    val songId: Int,
+    val songId: Long,
     val repeatMode: Int,
     val shuffleMode: Boolean,
     val currentPosition: Long
